refactor(enchantments): extract class infos to separate classes
